Split more parts and tests
