+++ public/app/features/explore/ExploreToolbar.tsx
@@ -11,7 +11,9 @@ import { DataQuery, RawTimeRange, TimeRange, TimeZone } from '@grafana/data';
 import { DataSourcePicker } from 'app/core/components/Select/DataSourcePicker';
 import { StoreState } from 'app/types/store';
 import { createAndCopyShortLink } from 'app/core/utils/shortLinks';
-import { changeDatasource, changeRefreshInterval, splitClose, splitOpen, syncTimes } from './state/actions';
+import { changeDatasource } from './state/datasource';
+import { splitClose, splitOpen } from './state/main';
+import { syncTimes, changeRefreshInterval } from './state/time';
 import { updateLocation } from 'app/core/actions';
 import { getTimeZone } from '../profile/state/selectors';
 import { updateTimeZoneForSession } from '../profile/state/reducers';
