+++ public/app/features/explore/state/exploreItem.ts
@@ -1,92 +1,227 @@
 import { AnyAction } from 'redux';
-import { DefaultTimeRange, LoadingState, LogsDedupStrategy, PanelData, sortLogsResult } from '@grafana/data';
-import { RefreshPicker } from '@grafana/ui';
 
-import { getQueryKeys, refreshIntervalToSortOrder, stopQueryState } from 'app/core/utils/explore';
-import { ExploreItemState, ExploreUpdateState } from 'app/types/explore';
-import {
-  changeDedupStrategyAction,
-  changeLoadingStateAction,
-  changeRangeAction,
-  changeRefreshIntervalAction,
-  changeSizeAction,
-  highlightLogsExpressionAction,
-  historyUpdatedAction,
-  initializeExploreAction,
-  loadDatasourceMissingAction,
-  loadDatasourcePendingAction,
-  loadDatasourceReadyAction,
-  queriesImportedAction,
-  scanStartAction,
-  scanStopAction,
-  setPausedStateAction,
-  setUrlReplacedAction,
-  toggleLogLevelAction,
-  updateDatasourceInstanceAction,
-} from './actionTypes';
-import { Emitter } from 'app/core/core';
+import { getQueryKeys } from 'app/core/utils/explore';
+import { ExploreId, ExploreItemState } from 'app/types/explore';
 import { queryReducer } from './query';
+import { datasourceReducer } from './datasource';
+import { timeReducer } from './time';
+import { historyReducer } from './history';
+import { makeExploreItemState, makeInitialUpdateState } from './utils';
+import { createAction, PayloadAction } from '@reduxjs/toolkit';
+import { DataQuery, ExploreUrlState, LogLevel, LogsDedupStrategy, TimeRange } from '@grafana/data';
+import { Emitter } from 'app/core/core';
+import {
+  clearQueryKeys,
+  ensureQueries,
+  generateNewKeyAndAddRefIdIfMissing,
+  getTimeRangeFromUrl,
+} from 'app/core/utils/explore';
+import { getRichHistory } from 'app/core/utils/richHistory';
+// Types
+import { ThunkResult } from 'app/types';
+import { getTimeZone } from 'app/features/profile/state/selectors';
+import { updateLocation } from '../../../core/actions';
+import { serializeStateToUrlParam } from '@grafana/data/src/utils/url';
+import { richHistoryUpdatedAction } from './main';
+import { runQueries, setQueriesAction } from './query';
+import { loadExploreDatasourcesAndSetDatasource } from './datasource';
+import { updateTime } from './time';
+import { toRawTimeRange } from '../utils/time';
 
-export const DEFAULT_RANGE = {
-  from: 'now-6h',
-  to: 'now',
+/**
+ * Keep track of the Explore container size, in particular the width.
+ * The width will be used to calculate graph intervals (number of datapoints).
+ */
+export interface ChangeSizePayload {
+  exploreId: ExploreId;
+  width: number;
+  height: number;
+}
+export const changeSizeAction = createAction<ChangeSizePayload>('explore/changeSize');
+
+/**
+ * Change deduplication strategy for logs.
+ */
+export interface ChangeDedupStrategyPayload {
+  exploreId: ExploreId;
+  dedupStrategy: LogsDedupStrategy;
+}
+export const changeDedupStrategyAction = createAction<ChangeDedupStrategyPayload>('explore/changeDedupStrategyAction');
+
+/**
+ * Highlight expressions in the log results
+ */
+export interface HighlightLogsExpressionPayload {
+  exploreId: ExploreId;
+  expressions: string[];
+}
+export const highlightLogsExpressionAction = createAction<HighlightLogsExpressionPayload>(
+  'explore/highlightLogsExpression'
+);
+
+/**
+ * Initialize Explore state with state from the URL and the React component.
+ * Call this only on components for with the Explore state has not been initialized.
+ */
+export interface InitializeExplorePayload {
+  exploreId: ExploreId;
+  containerWidth: number;
+  eventBridge: Emitter;
+  queries: DataQuery[];
+  range: TimeRange;
+  originPanelId?: number | null;
+}
+export const initializeExploreAction = createAction<InitializeExplorePayload>('explore/initializeExplore');
+
+export interface ToggleLogLevelPayload {
+  exploreId: ExploreId;
+  hiddenLogLevels: LogLevel[];
+}
+export const toggleLogLevelAction = createAction<ToggleLogLevelPayload>('explore/toggleLogLevel');
+
+export interface SetUrlReplacedPayload {
+  exploreId: ExploreId;
+}
+export const setUrlReplacedAction = createAction<SetUrlReplacedPayload>('explore/setUrlReplaced');
+
+/**
+ * Keep track of the Explore container size, in particular the width.
+ * The width will be used to calculate graph intervals (number of datapoints).
+ */
+export function changeSize(
+  exploreId: ExploreId,
+  { height, width }: { height: number; width: number }
+): PayloadAction<ChangeSizePayload> {
+  return changeSizeAction({ exploreId, height, width });
+}
+
+/**
+ * Change logs deduplication strategy.
+ */
+export const changeDedupStrategy = (
+  exploreId: ExploreId,
+  dedupStrategy: LogsDedupStrategy
+): PayloadAction<ChangeDedupStrategyPayload> => {
+  return changeDedupStrategyAction({ exploreId, dedupStrategy });
 };
 
-export const makeInitialUpdateState = (): ExploreUpdateState => ({
-  datasource: false,
-  queries: false,
-  range: false,
-  mode: false,
-});
+/**
+ * Initialize Explore state with state from the URL and the React component.
+ * Call this only on components for with the Explore state has not been initialized.
+ */
+export function initializeExplore(
+  exploreId: ExploreId,
+  datasourceName: string,
+  queries: DataQuery[],
+  range: TimeRange,
+  containerWidth: number,
+  eventBridge: Emitter,
+  originPanelId?: number | null
+): ThunkResult<void> {
+  return async (dispatch, getState) => {
+    dispatch(loadExploreDatasourcesAndSetDatasource(exploreId, datasourceName));
+    dispatch(
+      initializeExploreAction({
+        exploreId,
+        containerWidth,
+        eventBridge,
+        queries,
+        range,
+        originPanelId,
+      })
+    );
+    dispatch(updateTime({ exploreId }));
+    const richHistory = getRichHistory();
+    dispatch(richHistoryUpdatedAction({ richHistory }));
+  };
+}
 
 /**
- * Returns a fresh Explore area state
+ * Save local redux state back to the URL. Should be called when there is some change that should affect the URL.
+ * Not all of the redux state is reflected in URL though.
  */
-export const makeExploreItemState = (): ExploreItemState => ({
-  containerWidth: 0,
-  datasourceInstance: null,
-  requestedDatasourceName: null,
-  datasourceLoading: null,
-  datasourceMissing: false,
-  history: [],
-  queries: [],
-  initialized: false,
-  range: {
-    from: null,
-    to: null,
-    raw: DEFAULT_RANGE,
-  } as any,
-  absoluteRange: {
-    from: null,
-    to: null,
-  } as any,
-  scanning: false,
-  loading: false,
-  queryKeys: [],
-  urlState: null,
-  update: makeInitialUpdateState(),
-  latency: 0,
-  isLive: false,
-  isPaused: false,
-  urlReplaced: false,
-  queryResponse: createEmptyQueryResponse(),
-  tableResult: null,
-  graphResult: null,
-  logsResult: null,
-  dedupStrategy: LogsDedupStrategy.none,
-  eventBridge: (null as unknown) as Emitter,
-});
+export const stateSave = (): ThunkResult<void> => {
+  return (dispatch, getState) => {
+    const { left, right, split } = getState().explore;
+    const orgId = getState().user.orgId.toString();
+    const replace = left && left.urlReplaced === false;
+    const urlStates: { [index: string]: string } = { orgId };
+    const leftUrlState: ExploreUrlState = {
+      datasource: left.datasourceInstance!.name,
+      queries: left.queries.map(clearQueryKeys),
+      range: toRawTimeRange(left.range),
+    };
+    urlStates.left = serializeStateToUrlParam(leftUrlState, true);
+    if (split) {
+      const rightUrlState: ExploreUrlState = {
+        datasource: right.datasourceInstance!.name,
+        queries: right.queries.map(clearQueryKeys),
+        range: toRawTimeRange(right.range),
+      };
 
-export const createEmptyQueryResponse = (): PanelData => ({
-  state: LoadingState.NotStarted,
-  series: [],
-  timeRange: DefaultTimeRange,
-});
+      urlStates.right = serializeStateToUrlParam(rightUrlState, true);
+    }
+
+    dispatch(updateLocation({ query: urlStates, replace }));
+    if (replace) {
+      dispatch(setUrlReplacedAction({ exploreId: ExploreId.left }));
+    }
+  };
+};
 
 /**
- * Global Explore state that handles multiple Explore areas and the split state
+ * Reacts to changes in URL state that we need to sync back to our redux state. Checks the internal update variable
+ * to see which parts change and need to be synced.
+ * @param exploreId
  */
-export const initialExploreItemState = makeExploreItemState();
+export function refreshExplore(exploreId: ExploreId): ThunkResult<void> {
+  return (dispatch, getState) => {
+    const itemState = getState().explore[exploreId];
+    if (!itemState.initialized) {
+      return;
+    }
+
+    const { urlState, update, containerWidth, eventBridge } = itemState;
+
+    if (!urlState) {
+      return;
+    }
+
+    const { datasource, queries, range: urlRange, originPanelId } = urlState;
+    const refreshQueries: DataQuery[] = [];
+
+    for (let index = 0; index < queries.length; index++) {
+      const query = queries[index];
+      refreshQueries.push(generateNewKeyAndAddRefIdIfMissing(query, refreshQueries, index));
+    }
+
+    const timeZone = getTimeZone(getState().user);
+    const range = getTimeRangeFromUrl(urlRange, timeZone);
+
+    // need to refresh datasource
+    if (update.datasource) {
+      const initialQueries = ensureQueries(queries);
+      dispatch(
+        initializeExplore(exploreId, datasource, initialQueries, range, containerWidth, eventBridge, originPanelId)
+      );
+      return;
+    }
+
+    if (update.range) {
+      dispatch(updateTime({ exploreId, rawRange: range.raw }));
+    }
+
+    // need to refresh queries
+    if (update.queries) {
+      dispatch(setQueriesAction({ exploreId, queries: refreshQueries }));
+    }
+
+    // always run queries when refresh is needed
+    if (update.queries || update.range) {
+      dispatch(runQueries(exploreId));
+    }
+  };
+}
 
 /**
  * Reducer for an Explore area, to be used by the global Explore reducer.
@@ -98,34 +233,13 @@ export const initialExploreItemState = makeExploreItemState();
 // https://github.com/reduxjs/redux-toolkit/issues/242
 export const itemReducer = (state: ExploreItemState = makeExploreItemState(), action: AnyAction): ExploreItemState => {
   state = queryReducer(state, action);
+  state = datasourceReducer(state, action);
+  state = timeReducer(state, action);
+  state = historyReducer(state, action);
 
   if (changeSizeAction.match(action)) {
     const containerWidth = action.payload.width;
     return { ...state, containerWidth };
-  }
-
-  if (changeRefreshIntervalAction.match(action)) {
-    const { refreshInterval } = action.payload;
-    const live = RefreshPicker.isLive(refreshInterval);
-    const sortOrder = refreshIntervalToSortOrder(refreshInterval);
-    const logsResult = sortLogsResult(state.logsResult, sortOrder);
-
-    if (RefreshPicker.isLive(state.refreshInterval) && !live) {
-      stopQueryState(state.querySubscription);
-    }
-
-    return {
-      ...state,
-      refreshInterval,
-      queryResponse: {
-        ...state.queryResponse,
-        state: live ? LoadingState.Streaming : LoadingState.Done,
-      },
-      isLive: live,
-      isPaused: live ? false : state.isPaused,
-      loading: live,
-      logsResult,
-    };
   }
 
   if (highlightLogsExpressionAction.match(action)) {
@@ -156,77 +270,6 @@ export const itemReducer = (state: ExploreItemState = makeExploreItemState(), ac
     };
   }
 
-  if (updateDatasourceInstanceAction.match(action)) {
-    const { datasourceInstance } = action.payload;
-
-    // Custom components
-    stopQueryState(state.querySubscription);
-
-    return {
-      ...state,
-      datasourceInstance,
-      graphResult: null,
-      tableResult: null,
-      logsResult: null,
-      latency: 0,
-      queryResponse: createEmptyQueryResponse(),
-      loading: false,
-      queryKeys: [],
-      originPanelId: state.urlState && state.urlState.originPanelId,
-    };
-  }
-
-  if (loadDatasourceMissingAction.match(action)) {
-    return {
-      ...state,
-      datasourceMissing: true,
-      datasourceLoading: false,
-      update: makeInitialUpdateState(),
-    };
-  }
-
-  if (loadDatasourcePendingAction.match(action)) {
-    return {
-      ...state,
-      datasourceLoading: true,
-      requestedDatasourceName: action.payload.requestedDatasourceName,
-    };
-  }
-
-  if (loadDatasourceReadyAction.match(action)) {
-    const { history } = action.payload;
-    return {
-      ...state,
-      history,
-      datasourceLoading: false,
-      datasourceMissing: false,
-      logsHighlighterExpressions: undefined,
-      update: makeInitialUpdateState(),
-    };
-  }
-
-  if (scanStartAction.match(action)) {
-    return { ...state, scanning: true };
-  }
-
-  if (scanStopAction.match(action)) {
-    return {
-      ...state,
-      scanning: false,
-      scanRange: undefined,
-      update: makeInitialUpdateState(),
-    };
-  }
-
-  if (queriesImportedAction.match(action)) {
-    const { queries } = action.payload;
-    return {
-      ...state,
-      queries,
-      queryKeys: getQueryKeys(queries, state.datasourceInstance),
-    };
-  }
-
   if (toggleLogLevelAction.match(action)) {
     const { hiddenLogLevels } = action.payload;
     return {
@@ -235,47 +278,10 @@ export const itemReducer = (state: ExploreItemState = makeExploreItemState(), ac
     };
   }
 
-  if (historyUpdatedAction.match(action)) {
-    return {
-      ...state,
-      history: action.payload.history,
-    };
-  }
-
   if (setUrlReplacedAction.match(action)) {
     return {
       ...state,
       urlReplaced: true,
-    };
-  }
-
-  if (changeRangeAction.match(action)) {
-    const { range, absoluteRange } = action.payload;
-    return {
-      ...state,
-      range,
-      absoluteRange,
-      update: makeInitialUpdateState(),
-    };
-  }
-
-  if (changeLoadingStateAction.match(action)) {
-    const { loadingState } = action.payload;
-    return {
-      ...state,
-      queryResponse: {
-        ...state.queryResponse,
-        state: loadingState,
-      },
-      loading: loadingState === LoadingState.Loading || loadingState === LoadingState.Streaming,
-    };
-  }
-
-  if (setPausedStateAction.match(action)) {
-    const { isPaused } = action.payload;
-    return {
-      ...state,
-      isPaused: isPaused,
     };
   }
